尝试使用 Clean Architecture
