+++ catcove/__init__.py
@@ -1,1 +1,1 @@
-from .app import create_app
+from .web.app import create_app
