--- catcove/blueprints/api/dev/endpoints/about.py
@@ -1,17 +1,0 @@
-from sanic import Blueprint
-
-from ...response import model2json, code
-from .....models.schemas.response import APIResp
-from .....business.about import about_model
-
-api_about = Blueprint("about")
-
-@api_about.route("/about", error_format="json")
-async def about(request):
-    return model2json(
-        body=APIResp(
-            code=code.RETURNED_RESOURCE,
-            info="SUCCESS",
-            org=about_model()
-        )
-    )
