+++ catcove/dependencies/__init__.py
@@ -1,0 +1,25 @@
+from sanic import Sanic
+
+# Some contexts.
+from contextvars import ContextVar
+
+_base_model_session_ctx = ContextVar("session")
+
+
+def register_dependencies(app: Sanic) -> None:
+
+    # Database
+
+    # load session to request.
+    from .db import async_session
+
+    @app.on_request
+    async def inject_session(request):
+        request.ctx.session = async_session()
+        request.ctx.session_ctx_token = _base_model_session_ctx.set(request.ctx.session)
+    
+    @app.on_response
+    async def close_session(request, response):
+        if hasattr(request.ctx, "session_ctx_token"):
+            _base_model_session_ctx.reset(request.ctx.session_ctx_token)
+            await request.ctx.session.close()
