+++ catcove/web/blueprints/__init__.py
@@ -1,0 +1,10 @@
+from sanic import Sanic
+
+def register_routers(app: Sanic):
+    # API
+    from .api import api
+    app.blueprint(api)
+
+    # Views
+    # from .blueprints import views
+    # app.blueprint(views)
