+++ catcove/web/blueprints/api/dev/__init__.py
@@ -5,7 +5,7 @@ api_endpoint = Blueprint("api_dev")
 
 from ..response import code
 from ..response import model2json
-from ....models.schemas.response import APIResp
+from .....models.schemas.response import APIResp
 
 @api_endpoint.route("/", methods=["GET"], error_format="json")
 async def hello_api(request):
@@ -15,11 +15,8 @@ async def hello_api(request):
         org="Hello API!"
     ))
 
-from .endpoints.about import api_about
-
 api_dev = Blueprint.group(
     api_endpoint,
-    api_about,
     version=0.1,
     version_prefix="/api/v"
 )
