--- docs/api/code.md
@@ -1,11 +1,0 @@
-# 返回码
-
-| 返回代码 | 意义 | 出现环境 | 备注 |
-|:-------:|:----:|:-------:|:----:|
-|   0000  | 服务器出bog了 | 任何地方 |  |
-|   4399  | 
-|   4500  | 认证失败 | 认证时 |  |
-|   6000  | 来自客户端的数据解析错误 | 提交资源时 |  |
-|   6401  | 无所查询用户的结果 | 查找用户时 |  | 
-|   6002  | 来自客户端的JSON数据解析错误 | 提交资源时 |  |
-|   6700  | 成功获取要访问的资源 | 请求资源时 |  |
