--- docs/auth.md
@@ -1,65 +1,0 @@
-# 认证与鉴权
-
-确定访问以及使用网站用户的身份从来都是一个很严肃的问题，特别是在搭建具有社交性质的网站的情况下。如果访问并操作账号的用户不能被确定的话，很容易导致不可预料的后果。
-
-所以我们需要一种兼顾用户的方便以及占用尽量低的成本的方法，来认证用户。
-
----
-
-# 空想家
-
-## 认证（Authentication）
-
-1. 你能够收到邮箱，你就是账号的主人
-2. 你知道密码，你就是账号的主人
-
-### 流程
-
-- 你是人
-  - 能够使用网络
-  - 心智成熟
-- 你是账号的所有者
-  - 密码
-  - 通过邮件发送的令牌
-  - 密谜
-- 你不是某些不欢迎的人（Optional）
-  - 黑名单制度
-
-#### 人类验证
-
-验证码，或是初中数学题/高中语文题/专业课/常识
-
-#### 身份验证
-
-先注册再谈这个。
-
-##### 基于邮件
-
-略。
-
-##### 基于密谜
-
-不是秘密，是「用于密保的谜题」。
-
-## 给予其使用服务的权利（Authorization）
-
-> _「哼，本小姐才不是那些妖艳贱货呢」_
-
-已经验证了你是你，该怎么确定访问我的设备是你？
-
-### 客户端
-
-### 服务器端
-
----
-
-# 设计、实现与实操
-
-## 应用环境
-
-- 访问敏感信息（假设全站）
-- **作为用户**进行什么什么操作
-
-### Token
-
-采用JWT（AccessToken + RefreshToken），或存放在cookie里
