--- docs/interface.md
@@ -1,3 +1,0 @@
-# 连接与会话
-
-用户的客户端与服务器连接相关。
